Fix Provision options for EmbeddedTerraform Stack
The options column for a ServiceTerraformTemplate includes information
that isn't necessary for a provision request and the options sent to the
Stack.create_stack should be tailored to only include what is necessary.

Also this fixes the credential IDs being passed down since those values
were in the options.config_info.provision hash.